% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tables.R
\name{createDT}
\alias{createDT}
\title{Cria objeto datatable}
\usage{
createDT(
  data,
  fixed = 1,
  cols,
  formats,
  widths = c("400px", "200px", "200px"),
  align = "left",
  footer = T
)
}
\arguments{
\item{data}{data.frame com os resultados da query}

\item{fixed}{quantidade de colunas a esquerda que serao fixadas. So deve ser maior que
1 se tiverem mais de uma coluna de agrupamento. Nao pode ser menor que 1.}

\item{cols}{nome das colunas de valor que constituirao o datatable}

\item{formats}{formato das colunas de valor que irao compor o datatable}

\item{widths}{widths das primeiras colunas}

\item{align}{posicao para alinhar o datatable}

\item{footer}{T ou F}
}
\description{
Cria objeto datatable
}
