% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{filtrosEstrutura}
\alias{filtrosEstrutura}
\title{filtrosEstrutura}
\usage{
filtrosEstrutura(ncols, input_list)
}
\arguments{
\item{ncols}{numero de colunas a ser utilizado em cada linha. Deve ser um numero que divide 12. (1, 2, 3, 4, 6 ou 12).}

\item{input_list}{lista de inputs a serem posicionados, usualmente criado com cria_filtros.}
}
\description{
Funcao que monta a estrutura de colunas a partir dos filtros criados.
}
