% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_formata_dados.R
\name{formata_dados}
\alias{formata_dados}
\title{formata_dados}
\usage{
formata_dados(
  con,
  nometabela1,
  nometabela2,
  colunas_exploratorias,
  colunas_valor1,
  colunas_valor2
)
}
\arguments{
\item{con}{uma conexao com o SQL criada por `DBI::dbConnect`.}

\item{nometabela1, nometabela2}{nome das tabelas onde estão os dados.}

\item{colunas_exploratorias}{nome das colunas a serem usadas como filtro e/ou como tabela.}

\item{colunas_valor1, colunas_valor2}{colunas com os valores a serem exibidos nas tabelas, provinientes
de cada uma das tabelas (1 e 2)}
}
\value{
nao retorna nada, apenas cria as tabelas na conexao.
}
\description{
Funcao para auxiliar na formatacao necessaria para o app. Garante que as colunas
exploratórias existem nas duas tabelas e cria tabelas com as combinações únicas dessas colunas
para auxiliar nos inputs. Após isso, cria indexes em todas as tabelas para acelerar as queries.

Importante: as tabelas com nome {coluna_exploratoria}Id serão criadas no banco de dados, e se ja
existirem, serão dropadas.
}
