% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizacaoDados-package.R
\docType{package}
\name{VisualizacaoDados-package}
\alias{VisualizacaoDados}
\alias{VisualizacaoDados-package}
\title{VisualizacaoDados: Funções para modularizar e organizar a visualização de dados via Shiny}
\description{
Esse pacote fornece uma série de funções que criam um esqueleto robusto e escalável para construir visulizações de dados via Shiny.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vitorcapdeville/VisualizacaoDados}
  \item Report bugs at \url{https://github.com/vitorcapdeville/VisualizacaoDados/issues}
}

}
\author{
\strong{Maintainer}: Vitor Capdeville \email{vgcapdeville@hotmail.com}

}
\keyword{internal}
